Relocated all services files to src/service and fixed references to them
