+++ src/service/graph-generation/treeTraversal.js
