+++ src/service/pathway-commons/search/hgnc.js
